+++ client/cmd/sturdy/.gitignore
