+++ client/cmd/sturdy/build-and-upload.sh
