+++ client/cmd/sturdy/legal/legal.go
