+++ client/cmd/sturdy/msi/.gitignore
