+++ client/cmd/sturdy/msi/LICENSE
