+++ client/cmd/sturdy/msi/README.md
