+++ client/cmd/sturdy/msi/assets/main.wxs
