+++ client/cmd/sturdy/msi/assets/msi_banner.bmp
