+++ client/cmd/sturdy/msi/assets/msi_splash.bmp
