+++ client/cmd/sturdy/msi/assets/sturdy.ico
