+++ client/cmd/sturdy/msi/build.go
