+++ client/cmd/sturdy/msi/contents/sturdy.txt
