+++ client/cmd/sturdy/msi/helpers.go
