+++ client/cmd/sturdy/msi/run.ps1
