+++ client/cmd/sturdy/windows-installer.ps1
